kpbrandt/requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.3
django-rest==0.0.1
django-rest-swagger==2.1.2
djangorestframework==3.9.0
flake8==3.6.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mccabe==0.6.1
netifaces==0.10.7
openapi-codec==1.3.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
pyflakes==2.0.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.24.1
